Mursalim Haknazar Change course repo and course entity

+++ src/edu/aitu/oop3/entity/Course.java
@@ -1,17 +1,17 @@
 package edu.aitu.oop3.entity;
 
 public class Course {
-    private int id;
+    private Long id;
     private String title;
     private int capacity;
     private String schedule;
-    public Course(int id,String title, int capacity,String schedule){
+    public Course(Long id,String title, int capacity,String schedule){
         this.id=id;
         this.title=title;
         this.capacity=capacity;
         this.schedule=schedule;
     }
-    public int getid() {
+    public Long getid() {
         return id;
     }
     public String getTitle(){
