+++ src/edu/aitu/oop3/repository/CourseRepository.java
@@ -2,7 +2,7 @@ package edu.aitu.oop3.repository;
 import edu.aitu.oop3.entity.Course;
 import java.util.List;
 public interface CourseRepository {
-    void deleteByID(int id);
-    Course FindById(int id);
+    void deleteByID(Long id);
+    Course FindById(Long id);
     List<Course> findAll();
 }
